+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
-#### Version 0.14.0 (TBD)
+#### Version 0.14.1 (TBD)
+
+Fixed:
+* [#120](https://github.com/mailgun/kafka-pixy/pull/120) Consumption from a
+  topic stopped a group 
+
+#### Version 0.14.0 (2017-09-11)
 
 Implemented:
 * Added HTTP API endpoints to list topics optionally with partitions and